--- a/core/include/traccc/edm/measurement.hpp
+++ b/core/include/traccc/edm/measurement.hpp
@@ -48,8 +48,8 @@ struct measurement {
 TRACCC_HOST_DEVICE
 inline bool operator<(const measurement& lhs, const measurement& rhs) {
 
-    if (lhs.module_link != rhs.module_link) {
-        return lhs.module_link < rhs.module_link;
+    if (lhs.surface_link != rhs.surface_link) {
+        return lhs.surface_link < rhs.surface_link;
     } else if (std::abs(lhs.local[0] - rhs.local[0]) > float_epsilon) {
         return (lhs.local[0] < rhs.local[0]);
     } else {
@@ -61,12 +61,27 @@ inline bool operator<(const measurement& lhs, const measurement& rhs) {
 TRACCC_HOST_DEVICE
 inline bool operator==(const measurement& lhs, const measurement& rhs) {
 
-    return ((lhs.module_link == rhs.module_link) &&
+    return ((lhs.surface_link == rhs.surface_link) &&
             (std::abs(lhs.local[0] - rhs.local[0]) < float_epsilon) &&
             (std::abs(lhs.local[1] - rhs.local[1]) < float_epsilon) &&
             (std::abs(lhs.variance[0] - rhs.variance[0]) < float_epsilon) &&
             (std::abs(lhs.variance[1] - rhs.variance[1]) < float_epsilon));
 }
+
+/// Comparator based on detray barcode value
+struct measurement_sort_comp {
+    TRACCC_HOST_DEVICE
+    bool operator()(const measurement& lhs, const measurement& rhs) {
+        return lhs.surface_link < rhs.surface_link;
+    }
+};
+
+struct measurement_equal_comp {
+    TRACCC_HOST_DEVICE
+    bool operator()(const measurement& lhs, const measurement& rhs) {
+        return lhs.surface_link == rhs.surface_link;
+    }
+};
 
 /// Declare all measurement collection types
 using measurement_collection_types = collection_types<measurement>;