--- a/core/include/traccc/finding/finding_algorithm.ipp
+++ b/core/include/traccc/finding/finding_algorithm.ipp
@@ -19,28 +19,47 @@ namespace traccc {
 template <typename stepper_t, typename navigator_t>
 track_candidate_container_types::host
 finding_algorithm<stepper_t, navigator_t>::operator()(
-    const detector_type& det,
-    const measurement_container_types::host& measurements,
+    const detector_type& det, measurement_collection_types::host&& measurements,
     const bound_track_parameters_collection_types::host& seeds) const {
 
     track_candidate_container_types::host output_candidates;
 
-    /**********************
-     * Create module map
-     **********************/
+    /*****************************************************************
+     * Measurement Operations
+     *****************************************************************/
 
-    std::vector<std::pair<geometry_id, unsigned int>> module_map;
-    const int n_modules = measurements.size();
+    // Sort the measurements w.r.t geometry barcode
+    std::sort(measurements.begin(), measurements.end(),
+              measurement_sort_comp());
 
-    module_map.reserve(n_modules);
+    // Get copy of barcode uniques
+    std::vector<measurement> uniques;
+    uniques.resize(measurements.size());
 
-    for (int i = 0; i < n_modules; i++) {
-        module_map.push_back(
-            std::make_pair(measurements.at(i).header.module, i));
+    auto end = std::unique_copy(measurements.begin(), measurements.end(),
+                                uniques.begin(), measurement_equal_comp());
+    unsigned int n_modules = end - uniques.begin();
+
+    // Get upper bounds of unique elements
+    std::vector<unsigned int> upper_bounds;
+    upper_bounds.reserve(n_modules);
+    for (unsigned int i = 0; i < n_modules; i++) {
+        auto up = std::upper_bound(measurements.begin(), measurements.end(),
+                                   uniques[i], measurement_sort_comp());
+        upper_bounds.push_back(std::distance(measurements.begin(), up));
     }
 
-    // Sort module map to use it for binary search
-    std::sort(module_map.begin(), module_map.end());
+    // Get the number of measurements of each module
+    std::vector<unsigned int> sizes(n_modules);
+    std::adjacent_difference(upper_bounds.begin(), upper_bounds.end(),
+                             sizes.begin());
+
+    // Create barcode sequence
+    std::vector<detray::geometry::barcode> barcodes;
+    barcodes.reserve(n_modules);
+    for (unsigned int i = 0; i < n_modules; i++) {
+        barcodes.push_back(uniques[i].surface_link);
+    }
 
     /**********************
      * Find tracks
@@ -119,30 +138,33 @@ finding_algorithm<stepper_t, navigator_t>::operator()(
              * CKF
              *************************/
 
-            // Get module id
-            const auto module_id = in_param.surface_link();
+            // Get barcode and measurements range on surface
+            const auto bcd = in_param.surface_link();
+            std::pair<unsigned int, unsigned int> range;
 
-            unsigned int header_id;
-            if (std::binary_search(
-                    module_map.begin(), module_map.end(), module_id.value(),
-                    compare_pair_int<std::pair, unsigned int>())) {
-                const auto lo2 = std::lower_bound(
-                    module_map.begin(), module_map.end(), module_id.value(),
-                    compare_pair_int<std::pair, unsigned int>());
-                header_id = (*lo2).second;
+            // Find the corresponding index of bcd in barcode vector
+            unsigned int bcd_id;
+            if (std::binary_search(barcodes.begin(), barcodes.end(), bcd)) {
+                const auto lo2 =
+                    std::lower_bound(barcodes.begin(), barcodes.end(), bcd);
+                bcd_id = std::distance(barcodes.begin(), lo2);
+
+                if (lo2 == barcodes.begin()) {
+                    range.first = 0u;
+                    range.second = upper_bounds[bcd_id];
+                } else {
+                    range.first = upper_bounds[bcd_id - 1];
+                    range.second = upper_bounds[bcd_id];
+                }
             } else {
                 continue;
             }
 
-            // Get measurements on surface
-            const auto measurements_on_surface =
-                measurements.at(header_id).items;
-
-            const unsigned int n_meas = measurements_on_surface.size();
             unsigned int n_branches = 0;
 
             // Iterate over the measurements
-            for (unsigned int item_id = 0; item_id < n_meas; item_id++) {
+            for (unsigned int item_id = range.first; item_id < range.second;
+                 item_id++) {
                 if (n_branches > m_cfg.max_num_branches_per_surface) {
                     break;
                 }
@@ -150,7 +172,7 @@ finding_algorithm<stepper_t, navigator_t>::operator()(
                 bound_track_parameters bound_param(in_param.surface_link(),
                                                    in_param.vector(),
                                                    in_param.covariance());
-                const auto& meas = measurements_on_surface[item_id];
+                const auto& meas = measurements[item_id];
 
                 track_state<transform3_type> trk_state(meas);
 
@@ -170,9 +192,8 @@ finding_algorithm<stepper_t, navigator_t>::operator()(
 
                     n_branches++;
 
-                    links[step].push_back({{previous_step, in_param_id},
-                                           {header_id, item_id},
-                                           module_id});
+                    links[step].push_back(
+                        {{previous_step, in_param_id}, item_id});
 
                     /*********************************
                      * Propagate to the next surface
@@ -254,7 +275,7 @@ finding_algorithm<stepper_t, navigator_t>::operator()(
 
             auto& cand = *it;
 
-            cand = measurements.at(L.meas_link);
+            cand = measurements.at(L.meas_idx);
 
             // Break the loop if the iterator is at the first candidate and
             // fill the seed