--- a/device/common/include/traccc/finding/device/count_threads.hpp
+++ b/device/common/include/traccc/finding/device/count_threads.hpp
@@ -15,20 +15,25 @@ namespace traccc::device {
 /// Function evalulating the number of measurements to be iterated per thread
 /// and the total number of threads required
 ///
-/// @param[in] globalIndex          The index of the current thread
-/// @param[in] cfg                  Track finding config object
-/// @param[in] n_measurements_view  The number of measurements per track
-/// @param[in] n_total_measurements Total number of meausurments
-/// @param[out] n_threads_view      The number of threads per tracks
+/// @param[in] globalIndex           The index of the current thread
+/// @param[in] cfg                   Track finding config object
+/// @param[in] params_view           Input parameters view object
+/// @param[in] barcodes_view         Barcodes view object
+/// @param[in] sizes_view            The number of measurements of each module
+/// @param[in] n_in_params           The number of input track parameters
+/// @param[in] n_total_measurements  Total number of meausurments
+/// @param[out] n_threads_view       The number of threads per tracks
 /// @param[out] n_measurements_per_thread  The number of measurements to be
 /// iterated per thread
-/// @param[out] n_total_thread      Total number of threads
+/// @param[out] n_total_thread       Total number of threads
 ///
 template <typename config_t>
 TRACCC_DEVICE inline void count_threads(
     std::size_t globalIndex, const config_t cfg,
-    vecmem::data::vector_view<const unsigned int> n_measurements_view,
-    const unsigned int& n_total_measurements,
+    bound_track_parameters_collection_types::const_view params_view,
+    vecmem::data::vector_view<const detray::geometry::barcode> barcodes_view,
+    vecmem::data::vector_view<const unsigned int> sizes_view,
+    const int n_in_params, const int n_total_measurements,
     vecmem::data::vector_view<unsigned int> n_threads_view,
     unsigned int& n_measurements_per_thread, unsigned int& n_total_threads);