--- a/device/common/include/traccc/finding/device/impl/build_tracks.ipp
+++ b/device/common/include/traccc/finding/device/impl/build_tracks.ipp
@@ -11,7 +11,7 @@ namespace traccc::device {
 
 TRACCC_DEVICE inline void build_tracks(
     std::size_t globalIndex,
-    measurement_container_types::const_view measurements_view,
+    measurement_collection_types::const_view measurements_view,
     bound_track_parameters_collection_types::const_view seeds_view,
     vecmem::data::jagged_vector_view<const candidate_link> links_view,
     vecmem::data::jagged_vector_view<const unsigned int> param_to_link_view,
@@ -19,7 +19,7 @@ TRACCC_DEVICE inline void build_tracks(
         tips_view,
     track_candidate_container_types::view track_candidates_view) {
 
-    measurement_container_types::const_device measurements(measurements_view);
+    measurement_collection_types::const_device measurements(measurements_view);
 
     bound_track_parameters_collection_types::const_device seeds(seeds_view);
 
@@ -53,7 +53,7 @@ TRACCC_DEVICE inline void build_tracks(
          it++) {
 
         auto& cand = *it;
-        cand = {measurements.at(L.meas_link)};
+        cand = {measurements.at(L.meas_idx)};
 
         // Break the loop if the iterator is at the first candidate and fill the
         // seed