--- a/device/common/include/traccc/finding/device/impl/find_tracks.ipp
+++ b/device/common/include/traccc/finding/device/impl/find_tracks.ipp
@@ -19,9 +19,9 @@ template <typename detector_t, typename config_t>
 TRACCC_DEVICE inline void find_tracks(
     std::size_t globalIndex, const config_t cfg,
     typename detector_t::detector_view_type det_data,
-    measurement_container_types::const_view measurements_view,
-    vecmem::data::vector_view<const thrust::pair<geometry_id, unsigned int>>
-        module_map_view,
+    measurement_collection_types::const_view measurements_view,
+    vecmem::data::vector_view<const detray::geometry::barcode> barcodes_view,
+    vecmem::data::vector_view<const unsigned int> upper_bounds_view,
     bound_track_parameters_collection_types::const_view in_params_view,
     vecmem::data::vector_view<const unsigned int> n_threads_view,
     const unsigned int step, const unsigned int& n_measurements_per_thread,
@@ -38,11 +38,10 @@ TRACCC_DEVICE inline void find_tracks(
     detector_t det(det_data);
 
     // Measurement
-    measurement_container_types::const_device measurements(measurements_view);
-
-    // module map
-    vecmem::device_vector<const thrust::pair<geometry_id, unsigned int>>
-        module_map(module_map_view);
+    measurement_collection_types::const_device measurements(measurements_view);
+    vecmem::device_vector<const detray::geometry::barcode> barcodes(
+        barcodes_view);
+    vecmem::device_vector<const unsigned int> upper_bounds(upper_bounds_view);
 
     // Input parameters
     bound_track_parameters_collection_types::const_device in_params(
@@ -62,18 +61,7 @@ TRACCC_DEVICE inline void find_tracks(
                                          n_threads.end(), globalIndex + 1);
     const auto in_param_id = std::distance(n_threads.begin(), lo1);
 
-    // Get module id
-    const auto module_id = in_params.at(in_param_id).surface_link();
-
-    // Search for measuremetns header ID
-    const auto lo2 = thrust::lower_bound(
-        thrust::seq, module_map.begin(), module_map.end(), module_id.value(),
-        compare_pair_int<thrust::pair, unsigned int>());
-    const auto header_id = (*lo2).second;
-
     // Get measurements on surface
-    const auto measurements_on_surface = measurements.at(header_id).items;
-
     unsigned int ref;
     if (lo1 == n_threads.begin()) {
         ref = 0;
@@ -81,12 +69,29 @@ TRACCC_DEVICE inline void find_tracks(
         ref = *(lo1 - 1);
     }
 
+    // Get barcode
+    const auto bcd = in_params.at(in_param_id).surface_link();
+
+    const auto lo2 =
+        thrust::lower_bound(thrust::seq, barcodes.begin(), barcodes.end(), bcd);
+    const auto bcd_id = std::distance(barcodes.begin(), lo2);
+
+    // Find the range for measurements
+    std::pair<unsigned int, unsigned int> range;
+    if (lo2 == barcodes.begin()) {
+        range.first = 0u;
+        range.second = upper_bounds[bcd_id];
+    } else {
+        range.first = upper_bounds[bcd_id - 1];
+        range.second = upper_bounds[bcd_id];
+    }
+
     const unsigned int offset = globalIndex - ref;
     const unsigned int stride = offset * n_measurements_per_thread;
-    const unsigned int n_meas_on_surface = measurements_on_surface.size();
+    const unsigned int n_meas_on_surface = range.second - range.first;
 
     // Iterate over the measurements
-    const detray::surface<detector_t> sf{det, module_id};
+    const detray::surface<detector_t> sf{det, bcd};
 
     // Last step ID
     const unsigned int previous_step =
@@ -96,9 +101,10 @@ TRACCC_DEVICE inline void find_tracks(
         if (i + stride >= n_meas_on_surface) {
             break;
         }
+        const unsigned int meas_idx = i + stride + range.first;
 
         bound_track_parameters in_par = in_params.at(in_param_id);
-        const auto meas = measurements_on_surface.at(i + stride);
+        const auto meas = measurements.at(meas_idx);
         track_state<typename detector_t::transform3> trk_state(meas);
 
         // Run the Kalman update
@@ -117,9 +123,7 @@ TRACCC_DEVICE inline void find_tracks(
             const unsigned int l_pos = num_candidates.fetch_add(1);
 
             // @TODO; Consider max_num_branches_per_surface
-            links[l_pos] = {{previous_step, in_param_id},
-                            {header_id, i + stride},
-                            module_id};
+            links[l_pos] = {{previous_step, in_param_id}, meas_idx};
 
             out_params[l_pos] = trk_state.filtered();
         }