… device/common/CMakeLists.txt
… device/common/include/traccc/finding/device/count_measurements.hpp
… device/common/include/traccc/finding/device/impl/count_measurements.ipp
… device/common/include/traccc/finding/device/impl/make_module_map.ipp
… device/common/include/traccc/finding/device/make_barcode_sequence.hpp
… device/common/include/traccc/finding/device/make_module_map.hpp
… device/cuda/include/traccc/cuda/finding/finding_algorithm.hpp
… examples/run/cpu/seeding_example.cpp
… examples/run/cpu/truth_finding_example.cpp
… examples/run/cuda/seeding_example_cuda.cpp
… examples/run/cuda/truth_finding_example_cuda.cpp
… io/include/traccc/io/read_measurements.hpp
… io/src/csv/read_measurements.hpp
… io/src/read_measurements.cpp
… tests/cpu/test_ckf_sparse_tracks.cpp
… tests/cuda/test_ckf_sparse_tracks.cpp