### UPDATE

Now the measurement collection is sorted in finding_algorithm.

And here is the benchmark result that compares with the `main` branch

This PR:
```
==> Statistics ... 
- read    106619 spacepoints from 26364 modules
- created  (cpu)  113677 seeds
- created (cuda)  113667 seeds
- created  (cpu) 117725 found tracks
- created (cuda) 117678 found tracks
- created  (cpu) 117725 fitted tracks
- created (cuda) 117678 fitted tracks
==>Elapsed times...
            Hit reading  (cpu)  2596 ms
                Seeding (cuda)  16 ms
                Seeding  (cpu)  596 ms
           Track params (cuda)  1 ms
           Track params  (cpu)  22 ms
 Track finding with CKF (cuda)  1640 ms
  Track finding with CKF (cpu)  16201 ms
  Track fitting with KF (cuda)  864 ms
   Track fitting with KF (cpu)  10309 ms
                     Wall time  32269 ms
```

`main`:
```
==> Statistics ... 
- read    106619 spacepoints from 26364 modules
- created  (cpu)  113677 seeds
- created (cuda)  113668 seeds
- created  (cpu) 117725 found tracks
- created (cuda) 117678 found tracks
- created  (cpu) 117725 fitted tracks
- created (cuda) 117678 fitted tracks
==>Elapsed times...
            Hit reading  (cpu)  2878 ms
                Seeding (cuda)  15 ms
                Seeding  (cpu)  602 ms
           Track params (cuda)  2 ms
           Track params  (cpu)  23 ms
 Track finding with CKF (cuda)  1671 ms
  Track finding with CKF (cpu)  17029 ms
  Track fitting with KF (cuda)  948 ms
   Track fitting with KF (cpu)  10730 ms
                     Wall time  33918 ms
```
---------------------------------

This PR refactors CKF with the input of measurement collection.

Please note that the measurements should be sorted based on the detray::barcode before being passed to the CKF algorithm.

I will check the performance difference later but I don't think it is going to make a dramatic change ... :p