@beomki-yeo 2023-09-19
I guess getting const reference is a lot better for users as long as it doesn't take too much time for copy. Anyway it should be much cheaper than sort function