@beomki-yeo 2023-09-19
Just update on the benchmarking result.
I just realized that the nubmer of tracks after CUDA CKF is different. Maybe because the output of sort function is not determinisitc? 

This PR:

==> Statistics ... 
- read    106829 spacepoints from 26487 modules
- created  (cpu)  114031 seeds
- created (cuda)  114027 seeds
- created  (cpu) 117480 found tracks
- created (cuda) 117492 found tracks
- created  (cpu) 117480 fitted tracks
- created (cuda) 117492 fitted tracks
==>Elapsed times...
            Hit reading  (cpu)  2650 ms
                Seeding (cuda)  16 ms
                Seeding  (cpu)  605 ms
           Track params (cuda)  2 ms
           Track params  (cpu)  23 ms
 Track finding with CKF (cuda)  1741 ms
  Track finding with CKF (cpu)  18467 ms
  Track fitting with KF (cuda)  1117 ms
   Track fitting with KF (cpu)  11034 ms
                     Wall time  35678 ms

Main branch:

==> Statistics ... 
- read    106829 spacepoints from 26487 modules
- created  (cpu)  114031 seeds
- created (cuda)  114026 seeds
- created  (cpu) 117480 found tracks
- created (cuda) 117493 found tracks
- created  (cpu) 117480 fitted tracks
- created (cuda) 117493 fitted tracks
==>Elapsed times...
            Hit reading  (cpu)  2957 ms
                Seeding (cuda)  15 ms
                Seeding  (cpu)  604 ms
           Track params (cuda)  1 ms
           Track params  (cpu)  23 ms
 Track finding with CKF (cuda)  1679 ms
  Track finding with CKF (cpu)  18586 ms
  Track fitting with KF (cuda)  1213 ms
   Track fitting with KF (cpu)  11209 ms
                     Wall time  36308 ms